requirements.txt
Flask==1.0.4
requests==2.21.0
urllib3==1.24.2
apscheduler==3.6.3
connexion[swagger-ui]
pyodbc
pymongo
python-decouple
cryptography